+++ src/Quotinator.Core/Data/Enums/Genre.cs
@@ -1,0 +1,19 @@
+namespace Quotinator.Core.Data.Enums;
+
+public enum Genre
+{
+    Unknown = 0,
+    Action,
+    Adventure,
+    Animation,
+    Comedy,
+    Drama,
+    Fantasy,
+    Fiction,
+    Horror,
+    Mystery,
+    NonFiction,
+    Romance,
+    SciFi,
+    Thriller
+}