+++ src/Quotinator.Core/Data/Enums/QuoteType.cs
@@ -1,0 +1,11 @@
+namespace Quotinator.Core.Data.Enums;
+
+public enum QuoteType
+{
+    Unknown = 0,
+    Movie,
+    Tv,
+    Anime,
+    Book,
+    Person
+}