+++ src/Quotinator.Core/Data/IDbConnectionFactory.cs
@@ -1,0 +1,8 @@
+using System.Data;
+
+namespace Quotinator.Core.Data;
+
+public interface IDbConnectionFactory
+{
+    IDbConnection CreateConnection();
+}