+++ src/Quotinator.Core/Data/RecordBase.cs
@@ -1,0 +1,14 @@
+using Dapper.Contrib.Extensions;
+
+namespace Quotinator.Core.Data;
+
+public abstract class RecordBase
+{
+    [ExplicitKey]
+    public Guid Id { get; init; } = Guid.NewGuid();
+
+    public SafeValue<DateTime?> DateCreated  { get; init; } = SafeDateValue.Now;
+    public SafeValue<DateTime?> DateModified { get; set; }  = SafeDateValue.Empty;
+    public SafeValue<DateTime?> DateDeleted  { get; set; }  = SafeDateValue.Empty;
+    public bool IsDeleted { get; set; }
+}